[FLINK-5067] Java 8 build fixes
- New type inference rules changed which overloaded methods are picked.
- Renamed java.version property in maven to java.minor.version. This is due to the fact that overriding java.version is used by some third-party classes and they expect the version to be fully qualified.
- Introduced Java 8 build for Travis CI

+++ flink-tests/src/test/java/org/apache/flink/test/accumulators/AccumulatorIterativeITCase.java
@@ -48,7 +48,7 @@ public class AccumulatorIterativeITCase extends JavaProgramTestBase {
 		
 		iteration.closeWith(iteration.reduceGroup(new SumReducer())).output(new DiscardingOutputFormat<Integer>());
 		
-		Assert.assertEquals(NUM_ITERATIONS * 6, env.execute().getAccumulatorResult(ACC_NAME));
+		Assert.assertEquals(Integer.valueOf(NUM_ITERATIONS * 6), env.execute().getAccumulatorResult(ACC_NAME));
 	}
 	
 	static final class SumReducer extends RichGroupReduceFunction<Integer, Integer> {
